Video rewriting code is reusable now
