+++ src/apps/notion/models/video.py
@@ -1,9 +1,10 @@
 from django.utils.translation import gettext_lazy as _
 
 from core.models import TimestampedModel, models
+from core.video import VideoModelMixin
 
 
-class Video(TimestampedModel):
+class Video(TimestampedModel, VideoModelMixin):
     """Video mapping for multiple videohostings"""
 
     title = models.CharField(_("Title"), max_length=256, blank=True, null=True)
@@ -17,19 +18,3 @@ class Video(TimestampedModel):
 
     def __str__(self) -> str:
         return self.youtube_id
-
-    def get_youtube_embed_src(self) -> str:
-        return f"https://www.youtube.com/embed/{self.youtube_id}?rel=0"
-
-    def get_youtube_url(self) -> str:
-        return f"https://youtu.be/{self.youtube_id}"
-
-    def get_rutube_embed_src(self) -> str:
-        if self.rutube_access_key:
-            return f"https://rutube.ru/play/embed/{self.rutube_id}/?p={self.rutube_access_key}"
-        return f"https://rutube.ru/play/embed/{self.rutube_id}/"
-
-    def get_rutube_url(self) -> str:
-        if self.rutube_access_key:
-            return f"https://rutube.ru/video/{self.rutube_id}/?p={self.rutube_access_key}"
-        return f"https://rutube.ru/video/{self.rutube_id}/"
